Fix settings log tool filter
