+++ dashboard/src/components/logs.ts
@@ -22,6 +22,12 @@ import { createAsyncResource, loaded } from '../lib/async-state'
 import { toolCategory } from './tool-call-shared'
 import { StatusChip } from './common/status-chip'
 import {
+  detailLabel,
+  entryDetails,
+  logDisplayKind,
+  type LogDisplayKind,
+} from './log-classification'
+import {
   openIdeContextRouteLink,
   routeLinksForContext,
   type IdeContextRouteLink,
@@ -146,16 +152,6 @@ function categoryLabel(category: string | null | undefined): string | null {
   return CATEGORY_LABELS[category] ?? category
 }
 
-type LogDisplayKind =
-  | 'tool'
-  | 'turn'
-  | 'lifecycle'
-  | 'approval'
-  | 'broadcast'
-  | 'telemetry'
-  | 'task'
-  | 'log'
-
 const LOG_KIND_LABELS: Record<LogDisplayKind, string> = {
   tool: 'TOOL',
   turn: 'TURN',
@@ -232,47 +228,6 @@ export function deltaMergeCap(
 
 function sourceLabel(source: string): string {
   return SOURCE_LABELS[source] ?? (source || '(unknown source)')
-}
-
-function entryDetails(entry: LogEntry): Record<string, unknown> | null {
-  const details = entry.details
-  if (!details || typeof details !== 'object' || Array.isArray(details)) return null
-  return details
-}
-
-function detailLabel(details: Record<string, unknown> | null, key: string): string | null {
-  if (!details) return null
-  const value = details[key]
-  if (typeof value === 'string' && value.trim() !== '') return value.trim()
-  if (typeof value === 'number' && Number.isFinite(value)) return String(value)
-  return null
-}
-
-function logDisplayKind(entry: LogEntry): LogDisplayKind {
-  const details = entryDetails(entry)
-  const toolName = detailLabel(details, 'tool_name') ?? detailLabel(details, 'tool')
-  if (toolName) return 'tool'
-  switch (entry.category) {
-    case 'tool':
-      return 'tool'
-    case 'task':
-      return 'task'
-    case 'lifecycle':
-    case 'fsm':
-    case 'heartbeat':
-    case 'presence':
-      return 'lifecycle'
-    case 'directive':
-    case 'boundary':
-      return 'approval'
-    case 'telemetry':
-    case 'memory':
-      return 'telemetry'
-    case 'routine':
-      return entry.turn_id ? 'turn' : 'log'
-    default:
-      return entry.turn_id ? 'turn' : 'log'
-  }
 }
 
 function logSeverity(entry: LogEntry): 'ok' | 'warn' | 'bad' | 'busy' | 'info' {
